+++ src/AutoTable.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { polyfill } from "react-lifecycles-compat";
 import PropTypes from "prop-types";
 import { Data } from "perch-data";
-import { TableRow, TableCell, Typography, Checkbox } from "material-ui";
+import { TableRow, TableCell, Typography, Checkbox } from "@material-ui/core";
 import isEqual from "lodash.isequal";
 import { BaseTable as Table, LoadingRow } from "./";
 import { ActionButtonPropTypes } from "./ActionButton";
@@ -50,13 +50,13 @@ const getSortFromOrdering = ordering => {
 
 class AutoTable extends React.Component {
   static getDerivedStateFromProps(nextProps, prevState) {
-    if (!isEqual(nextProps.filter, prevState.filter)) {
+    if (!isEqual(nextProps.variables, prevState.variables)) {
       const { sortColumn, sortDirection } = getSortFromOrdering(
         nextProps.initialOrdering
       );
       return {
         allItems: new Set(),
-        filter: nextProps.filter,
+        variables: nextProps.variables,
         page: 1,
         sortColumn,
         sortDirection,
@@ -69,19 +69,19 @@ class AutoTable extends React.Component {
   constructor(props) {
     super(props);
 
-    const { filter, initialOrdering, rowsPerPage } = props;
+    const { initialOrdering, rowsPerPage, variables } = props;
     const { sortColumn, sortDirection } = getSortFromOrdering(initialOrdering);
 
     this.state = {
       allItems: new Set(),
-      filter,
       ordering: initialOrdering,
       page: 1,
       rowsPerPage,
       search: null,
       sortColumn,
       sortDirection,
-      selectedItems: new Set()
+      selectedItems: new Set(),
+      variables
     };
   }
 
@@ -123,7 +123,6 @@ class AutoTable extends React.Component {
     } = this.props;
     const { rowsPerPage, selectedItems } = this.state;
     const numColumns = multiselectable ? columns.length + 1 : columns.length;
-
     if (error) {
       return renderError ? (
         renderError(error, { columns, data, refetch, variables })
@@ -142,11 +141,16 @@ class AutoTable extends React.Component {
       );
     } else if (data) {
       this.setState({ allItems: new Set(data.results) });
-
-      return data.results.map(item => {
-        const cells = children(item, { columns, data, refetch, variables });
+      return data.results.map((item, index) => {
+        const cells = children(item, {
+          columns,
+          data,
+          index,
+          refetch,
+          variables
+        });
         return (
-          <TableRow hover key={item.id || JSON.stringify(item)}>
+          <TableRow hover key={item.id}>
             {multiselectable && (
               <TableCell padding="checkbox">
                 <Checkbox
@@ -218,7 +222,9 @@ class AutoTable extends React.Component {
   render() {
     const {
       action,
+      fullWidth,
       headerPadding,
+      options,
       pageable,
       searchable,
       sortable,
@@ -228,7 +234,7 @@ class AutoTable extends React.Component {
     } = this.props;
 
     const {
-      filter,
+      variables,
       ordering,
       page,
       rowsPerPage,
@@ -266,13 +272,13 @@ class AutoTable extends React.Component {
       }));
     }
 
-    const variables = { ...filter, ordering, page, rowsPerPage, search };
+    const dataVariables = { ...variables, ordering, page, rowsPerPage, search };
 
     const tableActions = multiselectActions.map(({ onClick, ...props }) => ({
       ...props,
       onClick: () => {
         this.setState({ selectedItems: new Set() });
-        return onClick([...selectedItems], { variables });
+        return onClick([...selectedItems], { variables: dataVariables });
       }
     }));
 
@@ -281,7 +287,7 @@ class AutoTable extends React.Component {
       : tableColumns;
 
     return (
-      <Data action={action} variables={variables}>
+      <Data action={action} options={options} variables={dataVariables}>
         {result => (
           <Table
             actions={tableActions}
@@ -290,6 +296,7 @@ class AutoTable extends React.Component {
                 ? tableColumns
                 : tableColumns.map(({ sortId, ...column }) => column)
             }
+            fullWidth={fullWidth}
             headerPadding={headerPadding}
             onSearch={this.handleSearch}
             onSort={this.handleSort}
@@ -303,7 +310,7 @@ class AutoTable extends React.Component {
           >
             {this.getTableBodyForResult(result, {
               columns: columnsWithoutCheckbox,
-              variables
+              variables: dataVariables
             })}
           </Table>
         )}
@@ -324,9 +331,12 @@ AutoTable.propTypes = {
       })
     ])
   ).isRequired,
-  filter: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  fullWidth: PropTypes.bool,
   headerPadding: PropTypes.string,
   initialOrdering: PropTypes.string,
+  multiselectable: PropTypes.bool,
+  multiselectActions: PropTypes.arrayOf(PropTypes.shape(ActionButtonPropTypes)),
+  options: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   pageable: PropTypes.bool,
   renderError: PropTypes.func,
   renderNoResults: PropTypes.func,
@@ -334,12 +344,11 @@ AutoTable.propTypes = {
   rowsPerPageOptions: PropTypes.arrayOf(PropTypes.number),
   searchable: PropTypes.bool,
   sortable: PropTypes.bool,
-  multiselectable: PropTypes.bool,
-  multiselectActions: PropTypes.arrayOf(PropTypes.shape(ActionButtonPropTypes))
+  variables: PropTypes.object // eslint-disable-line react/forbid-prop-types
 };
 
 AutoTable.defaultProps = {
-  filter: null,
+  fullWidth: false,
   headerPadding: "default",
   pageable: false,
   renderError: null,
@@ -350,7 +359,9 @@ AutoTable.defaultProps = {
   initialOrdering: null,
   sortable: false,
   multiselectable: false,
-  multiselectActions: []
+  multiselectActions: [],
+  options: null,
+  variables: null
 };
 
 polyfill(AutoTable);