+++ src/BaseTable.jsx
@@ -1,6 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Table, {
+import {
+  Hidden,
+  Table,
   TableBody,
   TableCell,
   TableFooter,
@@ -8,9 +10,8 @@ import Table, {
   TablePagination,
   TableSortLabel,
   TableRow
-} from "material-ui/Table";
-import { Hidden } from "material-ui";
-import { withStyles } from "material-ui/styles";
+} from "@material-ui/core";
+import { withStyles } from "@material-ui/core/styles";
 import { SearchBar, ActionBar } from "./";
 import { ActionButtonPropTypes } from "./ActionButton";
 
@@ -20,6 +21,9 @@ const toggleSort = (column, prevColumn, prevDirection) =>
   column === prevColumn ? toggleDirection(prevDirection) : "asc";
 
 const styles = () => ({
+  fullWidthContainer: {
+    width: "100%"
+  },
   headerCell: {
     whiteSpace: "nowrap",
     textTransform: "uppercase"
@@ -97,6 +101,7 @@ const BaseTable = ({
   children,
   classes,
   columns,
+  fullWidth,
   headerPadding,
   onSearch,
   onSort,
@@ -106,7 +111,7 @@ const BaseTable = ({
   sortColumn,
   sortDirection
 }) => (
-  <div>
+  <div className={fullWidth ? classes.fullWidthContainer : undefined}>
     {(searchable || actions) && (
       <div className={classes.topBar}>
         {searchable && (
@@ -207,6 +212,7 @@ BaseTable.propTypes = {
       })
     ])
   ).isRequired,
+  fullWidth: PropTypes.bool,
   headerPadding: PropTypes.string,
   onSearch: PropTypes.func,
   onSort: PropTypes.func,
@@ -227,6 +233,7 @@ BaseTable.propTypes = {
 BaseTable.defaultProps = {
   actions: null,
   children: [],
+  fullWidth: false,
   headerPadding: "default",
   onSearch: null,
   onSort: null,